server/requirements.txt
flask
Flask-Migrate 
Flask-SQLAlchemy
psycopg2
werkzeug
psycopg2-binary
flask-jwt-extended
python-dotenv
flask-WTF
Flask-SQLAlchemy
pycryptodome
cryptography 